public / private 間違い修正
