+++ Source/PLATEAURuntime/Public/RoadAdjust/PLATEAUReproducedRoad.h
@@ -18,6 +18,10 @@ enum class EPLATEAURoadLineType : uint8 {
     Crossing = 4 UMETA(DisplayName = "Crossing"),
 };
 
+/**
+* @brief Road Line 生成用パラメータ
+*
+*/
 USTRUCT(BlueprintType)
 struct PLATEAURUNTIME_API FPLATEAURoadLineParam {
     GENERATED_BODY()
@@ -31,12 +35,15 @@ struct PLATEAURUNTIME_API FPLATEAURoadLineParam {
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "PLATEAU|RoadAdjust")
     UMaterialInterface* LineMaterial;
 
+    //Mesh間の隙間
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "PLATEAU|RoadAdjust")
     float LineGap;
 
+    //元メッシュの幅に対する倍率
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "PLATEAU|RoadAdjust")
     float LineXScale;
 
+    //この値が0の場合アセットMeshのBoundsのサイズを利用
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "PLATEAU|RoadAdjust")
     float LineLength;
 };
@@ -56,15 +63,14 @@ protected:
     // Called when the game starts or when spawneds
     virtual void BeginPlay() override;
 
-    void GetVectors(TArray<FVector>& Vectors, URnWay* Way) const;
-
-    void CreateLineComponentByType(EPLATEAURoadLineType type, TArray<FVector> LinePoints, int32 index);
-
 public:
     virtual void Tick(float DeltaTime) override;
 
 private:
 
+    int32 NumComponents;
     TMap<EPLATEAURoadLineType, FPLATEAURoadLineParam> LineTypeMap;
 
+    void CreateLineTypeMap();
+    void CreateLineComponentByType(EPLATEAURoadLineType Type, TArray<FVector> LinePoints);
 };
